LibAwsSdkutils

Documentation for LibAwsSdkutils.

LibAwsSdkutils.AWS_C_SDKUTILS_PACKAGE_ID

LibAwsSdkutils.aws_endpoints_parameter

LibAwsSdkutils.aws_endpoints_parameter_type

LibAwsSdkutils.aws_endpoints_request_context

LibAwsSdkutils.aws_endpoints_resolved_endpoint

LibAwsSdkutils.aws_endpoints_resolved_endpoint_type

LibAwsSdkutils.aws_endpoints_rule_engine

LibAwsSdkutils.aws_endpoints_ruleset

LibAwsSdkutils.aws_partitions_config

LibAwsSdkutils.aws_profile

LibAwsSdkutils.aws_profile_collection

LibAwsSdkutils.aws_profile_property

LibAwsSdkutils.aws_profile_section_type

LibAwsSdkutils.aws_profile_source_type

LibAwsSdkutils.aws_resource_name

LibAwsSdkutils.aws_sdkutils_errors

LibAwsSdkutils.aws_sdkutils_log_subject

LibAwsSdkutils.aws_byte_buf_append_resource_name

LibAwsSdkutils.aws_endpoints_get_supported_ruleset_version

LibAwsSdkutils.aws_endpoints_parameter_get_built_in

LibAwsSdkutils.aws_endpoints_parameter_get_default_boolean

LibAwsSdkutils.aws_endpoints_parameter_get_default_string

LibAwsSdkutils.aws_endpoints_parameter_get_deprecated_message

LibAwsSdkutils.aws_endpoints_parameter_get_deprecated_since

LibAwsSdkutils.aws_endpoints_parameter_get_documentation

LibAwsSdkutils.aws_endpoints_parameter_get_is_required

LibAwsSdkutils.aws_endpoints_parameter_get_type

LibAwsSdkutils.aws_endpoints_parameters_get_is_deprecated

LibAwsSdkutils.aws_endpoints_request_context_acquire

LibAwsSdkutils.aws_endpoints_request_context_add_boolean

LibAwsSdkutils.aws_endpoints_request_context_add_string

LibAwsSdkutils.aws_endpoints_request_context_new

LibAwsSdkutils.aws_endpoints_request_context_release

LibAwsSdkutils.aws_endpoints_resolved_endpoint_acquire

LibAwsSdkutils.aws_endpoints_resolved_endpoint_get_error

LibAwsSdkutils.aws_endpoints_resolved_endpoint_get_headers

LibAwsSdkutils.aws_endpoints_resolved_endpoint_get_properties

LibAwsSdkutils.aws_endpoints_resolved_endpoint_get_type

LibAwsSdkutils.aws_endpoints_resolved_endpoint_get_url

LibAwsSdkutils.aws_endpoints_resolved_endpoint_release

LibAwsSdkutils.aws_endpoints_rule_engine_acquire

LibAwsSdkutils.aws_endpoints_rule_engine_new

LibAwsSdkutils.aws_endpoints_rule_engine_release

LibAwsSdkutils.aws_endpoints_rule_engine_resolve

LibAwsSdkutils.aws_endpoints_ruleset_acquire

LibAwsSdkutils.aws_endpoints_ruleset_get_parameters

LibAwsSdkutils.aws_endpoints_ruleset_get_service_id

LibAwsSdkutils.aws_endpoints_ruleset_get_version

LibAwsSdkutils.aws_endpoints_ruleset_new_from_string

LibAwsSdkutils.aws_endpoints_ruleset_release

LibAwsSdkutils.aws_get_config_file_path

LibAwsSdkutils.aws_get_credentials_file_path

LibAwsSdkutils.aws_get_profile_name

LibAwsSdkutils.aws_partitions_config_acquire

LibAwsSdkutils.aws_partitions_config_new_from_string

LibAwsSdkutils.aws_partitions_config_release

LibAwsSdkutils.aws_partitions_get_supported_version

LibAwsSdkutils.aws_profile_collection_acquire

LibAwsSdkutils.aws_profile_collection_destroy

LibAwsSdkutils.aws_profile_collection_get_profile

LibAwsSdkutils.aws_profile_collection_get_profile_count

LibAwsSdkutils.aws_profile_collection_get_section

LibAwsSdkutils.aws_profile_collection_get_section_count

LibAwsSdkutils.aws_profile_collection_new_from_buffer

LibAwsSdkutils.aws_profile_collection_new_from_file

LibAwsSdkutils.aws_profile_collection_new_from_merge

LibAwsSdkutils.aws_profile_collection_release

LibAwsSdkutils.aws_profile_get_name

LibAwsSdkutils.aws_profile_get_property

LibAwsSdkutils.aws_profile_get_property_count

LibAwsSdkutils.aws_profile_property_get_sub_property

LibAwsSdkutils.aws_profile_property_get_sub_property_count

LibAwsSdkutils.aws_profile_property_get_value

LibAwsSdkutils.aws_resource_name_init_from_cur

LibAwsSdkutils.aws_resource_name_length

LibAwsSdkutils.aws_sdkutils_library_clean_up

LibAwsSdkutils.aws_sdkutils_library_init

LibAwsSdkutils.AWS_C_SDKUTILS_PACKAGE_ID — Constant

Documentation not found.

LibAwsSdkutils.aws_endpoints_parameter — Type

Documentation not found.

LibAwsSdkutils.aws_endpoints_parameter_type — Type

aws_endpoints_parameter_type

Documentation not found.

LibAwsSdkutils.aws_endpoints_request_context — Type

Documentation not found.

LibAwsSdkutils.aws_endpoints_resolved_endpoint — Type

Documentation not found.

LibAwsSdkutils.aws_endpoints_resolved_endpoint_type — Type

aws_endpoints_resolved_endpoint_type

Documentation not found.

LibAwsSdkutils.aws_endpoints_rule_engine — Type

Documentation not found.

LibAwsSdkutils.aws_endpoints_ruleset — Type

Documentation not found.

LibAwsSdkutils.aws_partitions_config — Type

Documentation not found.

LibAwsSdkutils.aws_profile — Type

Documentation not found.

LibAwsSdkutils.aws_profile_collection — Type

Documentation not found.

LibAwsSdkutils.aws_profile_property — Type

Documentation not found.

LibAwsSdkutils.aws_profile_section_type — Type

aws_profile_section_type

Documentation not found.

LibAwsSdkutils.aws_profile_source_type — Type

aws_profile_source_type

The profile specification has rule exceptions based on what file the profile collection comes from.

LibAwsSdkutils.aws_resource_name — Type

aws_resource_name

Documentation not found.

LibAwsSdkutils.aws_sdkutils_errors — Type

aws_sdkutils_errors

Documentation not found.

LibAwsSdkutils.aws_sdkutils_log_subject — Type

aws_sdkutils_log_subject

Documentation not found.

LibAwsSdkutils.aws_byte_buf_append_resource_name — Method

aws_byte_buf_append_resource_name(buf, arn)

Serializes an ARN structure into the lexical string format

Prototype

int aws_byte_buf_append_resource_name(struct aws_byte_buf *buf, const struct aws_resource_name *arn);

LibAwsSdkutils.aws_endpoints_get_supported_ruleset_version — Method

aws_endpoints_get_supported_ruleset_version()

Documentation not found.

Prototype

struct aws_byte_cursor aws_endpoints_get_supported_ruleset_version(void);

LibAwsSdkutils.aws_endpoints_parameter_get_built_in — Method

aws_endpoints_parameter_get_built_in(parameter)

Documentation not found.

Prototype

struct aws_byte_cursor aws_endpoints_parameter_get_built_in( const struct aws_endpoints_parameter *parameter);

LibAwsSdkutils.aws_endpoints_parameter_get_default_boolean — Method

aws_endpoints_parameter_get_default_boolean(parameter, out_bool)

Documentation not found.

Prototype

int aws_endpoints_parameter_get_default_boolean( const struct aws_endpoints_parameter *parameter, const bool **out_bool);

LibAwsSdkutils.aws_endpoints_parameter_get_default_string — Method

aws_endpoints_parameter_get_default_string(parameter, out_cursor)

Documentation not found.

Prototype

int aws_endpoints_parameter_get_default_string( const struct aws_endpoints_parameter *parameter, struct aws_byte_cursor *out_cursor);

LibAwsSdkutils.aws_endpoints_parameter_get_deprecated_message — Method

aws_endpoints_parameter_get_deprecated_message(parameter)

Documentation not found.

Prototype

struct aws_byte_cursor aws_endpoints_parameter_get_deprecated_message( const struct aws_endpoints_parameter *parameter);

LibAwsSdkutils.aws_endpoints_parameter_get_deprecated_since — Method

aws_endpoints_parameter_get_deprecated_since(parameter)

Documentation not found.

Prototype

struct aws_byte_cursor aws_endpoints_parameter_get_deprecated_since( const struct aws_endpoints_parameter *parameter);

LibAwsSdkutils.aws_endpoints_parameter_get_documentation — Method

aws_endpoints_parameter_get_documentation(parameter)

Documentation not found.

Prototype

struct aws_byte_cursor aws_endpoints_parameter_get_documentation( const struct aws_endpoints_parameter *parameter);

LibAwsSdkutils.aws_endpoints_parameter_get_is_required — Method

aws_endpoints_parameter_get_is_required(parameter)

Documentation not found.

Prototype

bool aws_endpoints_parameter_get_is_required(const struct aws_endpoints_parameter *parameter);

LibAwsSdkutils.aws_endpoints_parameter_get_type — Method

aws_endpoints_parameter_get_type(parameter)

Documentation not found.

Prototype

enum aws_endpoints_parameter_type aws_endpoints_parameter_get_type( const struct aws_endpoints_parameter *parameter);

LibAwsSdkutils.aws_endpoints_parameters_get_is_deprecated — Method

aws_endpoints_parameters_get_is_deprecated(parameter)

Documentation not found.

Prototype

bool aws_endpoints_parameters_get_is_deprecated(const struct aws_endpoints_parameter *parameter);

LibAwsSdkutils.aws_endpoints_request_context_acquire — Method

aws_endpoints_request_context_acquire(request_context)

Documentation not found.

Prototype

struct aws_endpoints_request_context *aws_endpoints_request_context_acquire( struct aws_endpoints_request_context *request_context);

LibAwsSdkutils.aws_endpoints_request_context_add_boolean — Method

aws_endpoints_request_context_add_boolean(allocator, context, name, value)

Documentation not found.

Prototype

int aws_endpoints_request_context_add_boolean( struct aws_allocator *allocator, struct aws_endpoints_request_context *context, struct aws_byte_cursor name, bool value);

LibAwsSdkutils.aws_endpoints_request_context_add_string — Method

aws_endpoints_request_context_add_string(allocator, context, name, value)

Documentation not found.

Prototype

int aws_endpoints_request_context_add_string( struct aws_allocator *allocator, struct aws_endpoints_request_context *context, struct aws_byte_cursor name, struct aws_byte_cursor value);

LibAwsSdkutils.aws_endpoints_request_context_new — Method

aws_endpoints_request_context_new(allocator)

Documentation not found.

Prototype

struct aws_endpoints_request_context *aws_endpoints_request_context_new( struct aws_allocator *allocator);

LibAwsSdkutils.aws_endpoints_request_context_release — Method

aws_endpoints_request_context_release(request_context)

Documentation not found.

Prototype

struct aws_endpoints_request_context *aws_endpoints_request_context_release( struct aws_endpoints_request_context *request_context);

LibAwsSdkutils.aws_endpoints_resolved_endpoint_acquire — Method

aws_endpoints_resolved_endpoint_acquire(resolved_endpoint)

Documentation not found.

Prototype

struct aws_endpoints_resolved_endpoint *aws_endpoints_resolved_endpoint_acquire( struct aws_endpoints_resolved_endpoint *resolved_endpoint);

LibAwsSdkutils.aws_endpoints_resolved_endpoint_get_error — Method

aws_endpoints_resolved_endpoint_get_error(resolved_endpoint, out_error)

Documentation not found.

Prototype

int aws_endpoints_resolved_endpoint_get_error( const struct aws_endpoints_resolved_endpoint *resolved_endpoint, struct aws_byte_cursor *out_error);

LibAwsSdkutils.aws_endpoints_resolved_endpoint_get_headers — Method

aws_endpoints_resolved_endpoint_get_headers(resolved_endpoint, out_headers)

Documentation not found.

Prototype

int aws_endpoints_resolved_endpoint_get_headers( const struct aws_endpoints_resolved_endpoint *resolved_endpoint, const struct aws_hash_table **out_headers);

LibAwsSdkutils.aws_endpoints_resolved_endpoint_get_properties — Method

aws_endpoints_resolved_endpoint_get_properties(resolved_endpoint, out_properties)

Documentation not found.

Prototype

int aws_endpoints_resolved_endpoint_get_properties( const struct aws_endpoints_resolved_endpoint *resolved_endpoint, struct aws_byte_cursor *out_properties);

LibAwsSdkutils.aws_endpoints_resolved_endpoint_get_type — Method

aws_endpoints_resolved_endpoint_get_type(resolved_endpoint)

Documentation not found.

Prototype

enum aws_endpoints_resolved_endpoint_type aws_endpoints_resolved_endpoint_get_type( const struct aws_endpoints_resolved_endpoint *resolved_endpoint);

LibAwsSdkutils.aws_endpoints_resolved_endpoint_get_url — Method

aws_endpoints_resolved_endpoint_get_url(resolved_endpoint, out_url)

Documentation not found.

Prototype

int aws_endpoints_resolved_endpoint_get_url( const struct aws_endpoints_resolved_endpoint *resolved_endpoint, struct aws_byte_cursor *out_url);

LibAwsSdkutils.aws_endpoints_resolved_endpoint_release — Method

aws_endpoints_resolved_endpoint_release(resolved_endpoint)

Documentation not found.

Prototype

struct aws_endpoints_resolved_endpoint *aws_endpoints_resolved_endpoint_release( struct aws_endpoints_resolved_endpoint *resolved_endpoint);

LibAwsSdkutils.aws_endpoints_rule_engine_acquire — Method

aws_endpoints_rule_engine_acquire(rule_engine)

Documentation not found.

Prototype

struct aws_endpoints_rule_engine *aws_endpoints_rule_engine_acquire( struct aws_endpoints_rule_engine *rule_engine);

LibAwsSdkutils.aws_endpoints_rule_engine_new — Method

aws_endpoints_rule_engine_new(allocator, ruleset, partitions_config)

Create new rule engine for a given ruleset. In cases of failure NULL is returned and last error is set.

Prototype

struct aws_endpoints_rule_engine *aws_endpoints_rule_engine_new( struct aws_allocator *allocator, struct aws_endpoints_ruleset *ruleset, struct aws_partitions_config *partitions_config);

LibAwsSdkutils.aws_endpoints_rule_engine_release — Method

aws_endpoints_rule_engine_release(rule_engine)

Documentation not found.

Prototype

struct aws_endpoints_rule_engine *aws_endpoints_rule_engine_release( struct aws_endpoints_rule_engine *rule_engine);

LibAwsSdkutils.aws_endpoints_rule_engine_resolve — Method

aws_endpoints_rule_engine_resolve(engine, context, out_resolved_endpoint)

Documentation not found.

Prototype

int aws_endpoints_rule_engine_resolve( struct aws_endpoints_rule_engine *engine, const struct aws_endpoints_request_context *context, struct aws_endpoints_resolved_endpoint **out_resolved_endpoint);

LibAwsSdkutils.aws_endpoints_ruleset_acquire — Method

aws_endpoints_ruleset_acquire(ruleset)

Documentation not found.

Prototype

struct aws_endpoints_ruleset *aws_endpoints_ruleset_acquire(struct aws_endpoints_ruleset *ruleset);

LibAwsSdkutils.aws_endpoints_ruleset_get_parameters — Method

aws_endpoints_ruleset_get_parameters(ruleset)

Documentation not found.

Prototype

const struct aws_hash_table *aws_endpoints_ruleset_get_parameters( struct aws_endpoints_ruleset *ruleset);

LibAwsSdkutils.aws_endpoints_ruleset_get_service_id — Method

aws_endpoints_ruleset_get_service_id(ruleset)

Documentation not found.

Prototype

struct aws_byte_cursor aws_endpoints_ruleset_get_service_id( const struct aws_endpoints_ruleset *ruleset);

LibAwsSdkutils.aws_endpoints_ruleset_get_version — Method

aws_endpoints_ruleset_get_version(ruleset)

Documentation not found.

Prototype

struct aws_byte_cursor aws_endpoints_ruleset_get_version(const struct aws_endpoints_ruleset *ruleset);

LibAwsSdkutils.aws_endpoints_ruleset_new_from_string — Method

aws_endpoints_ruleset_new_from_string(allocator, ruleset_json)

Documentation not found.

Prototype

struct aws_endpoints_ruleset *aws_endpoints_ruleset_new_from_string( struct aws_allocator *allocator, struct aws_byte_cursor ruleset_json);

LibAwsSdkutils.aws_endpoints_ruleset_release — Method

aws_endpoints_ruleset_release(ruleset)

Documentation not found.

Prototype

struct aws_endpoints_ruleset *aws_endpoints_ruleset_release(struct aws_endpoints_ruleset *ruleset);

LibAwsSdkutils.aws_get_config_file_path — Method

aws_get_config_file_path(allocator, override_path)

Computes the final platform-specific path for the profile config file. Does limited home directory expansion/resolution.

override_path, if not null, will be searched first instead of using the standard home directory config path

Prototype

struct aws_string *aws_get_config_file_path( struct aws_allocator *allocator, const struct aws_byte_cursor *override_path);

LibAwsSdkutils.aws_get_credentials_file_path — Method

aws_get_credentials_file_path(allocator, override_path)

Computes the final platform-specific path for the profile credentials file. Does limited home directory expansion/resolution.

override_path, if not null, will be searched first instead of using the standard home directory config path

Prototype

struct aws_string *aws_get_credentials_file_path( struct aws_allocator *allocator, const struct aws_byte_cursor *override_path);

LibAwsSdkutils.aws_get_profile_name — Method

aws_get_profile_name(allocator, override_name)

Computes the profile to use for credentials lookups based on profile resolution rules

Prototype

struct aws_string *aws_get_profile_name(struct aws_allocator *allocator, const struct aws_byte_cursor *override_name);

LibAwsSdkutils.aws_partitions_config_acquire — Method

aws_partitions_config_acquire(partitions)

Documentation not found.

Prototype

struct aws_partitions_config *aws_partitions_config_acquire(struct aws_partitions_config *partitions);

LibAwsSdkutils.aws_partitions_config_new_from_string — Method

aws_partitions_config_new_from_string(allocator, json)

Documentation not found.

Prototype

struct aws_partitions_config *aws_partitions_config_new_from_string( struct aws_allocator *allocator, struct aws_byte_cursor json);

LibAwsSdkutils.aws_partitions_config_release — Method

aws_partitions_config_release(partitions)

Documentation not found.

Prototype

struct aws_partitions_config *aws_partitions_config_release(struct aws_partitions_config *partitions);

LibAwsSdkutils.aws_partitions_get_supported_version — Method

aws_partitions_get_supported_version()

Documentation not found.

Prototype

struct aws_byte_cursor aws_partitions_get_supported_version(void);

LibAwsSdkutils.aws_profile_collection_acquire — Method

aws_profile_collection_acquire(collection)

Increments the reference count on the profile collection, allowing the caller to take a reference to it.

Returns the same profile collection passed in.

Prototype

struct aws_profile_collection *aws_profile_collection_acquire(struct aws_profile_collection *collection);

LibAwsSdkutils.aws_profile_collection_destroy — Method

aws_profile_collection_destroy(profile_collection)

Deprecated

This is equivalent to aws_profile_collection_release.

Prototype

void aws_profile_collection_destroy(struct aws_profile_collection *profile_collection);

LibAwsSdkutils.aws_profile_collection_get_profile — Method

aws_profile_collection_get_profile(profile_collection, profile_name)

Retrieves a reference to a profile with the specified name, if it exists, from the profile collection

Prototype

const struct aws_profile *aws_profile_collection_get_profile( const struct aws_profile_collection *profile_collection, const struct aws_string *profile_name);

LibAwsSdkutils.aws_profile_collection_get_profile_count — Method

aws_profile_collection_get_profile_count(profile_collection)

Returns the number of profiles in a collection

Prototype

size_t aws_profile_collection_get_profile_count(const struct aws_profile_collection *profile_collection);

LibAwsSdkutils.aws_profile_collection_get_section — Method

aws_profile_collection_get_section(profile_collection, section_type, section_name)

Documentation not found.

Prototype

const struct aws_profile *aws_profile_collection_get_section( const struct aws_profile_collection *profile_collection, const enum aws_profile_section_type section_type, const struct aws_string *section_name);

LibAwsSdkutils.aws_profile_collection_get_section_count — Method

aws_profile_collection_get_section_count(profile_collection, section_type)

Returns the number of elements of the specified section in a collection.

Prototype

size_t aws_profile_collection_get_section_count( const struct aws_profile_collection *profile_collection, const enum aws_profile_section_type section_type);

LibAwsSdkutils.aws_profile_collection_new_from_buffer — Method

aws_profile_collection_new_from_buffer(allocator, buffer, source)

Create a new profile collection by parsing text in a buffer. Primarily for testing.

Prototype

struct aws_profile_collection *aws_profile_collection_new_from_buffer( struct aws_allocator *allocator, const struct aws_byte_buf *buffer, enum aws_profile_source_type source);

LibAwsSdkutils.aws_profile_collection_new_from_file — Method

aws_profile_collection_new_from_file(allocator, file_path, source)

Create a new profile collection by parsing a file with the specified path

Prototype

struct aws_profile_collection *aws_profile_collection_new_from_file( struct aws_allocator *allocator, const struct aws_string *file_path, enum aws_profile_source_type source);

LibAwsSdkutils.aws_profile_collection_new_from_merge — Method

aws_profile_collection_new_from_merge(allocator, config_profiles, credentials_profiles)

Create a new profile collection by merging a config-file-based profile collection and a credentials-file-based profile collection

Prototype

struct aws_profile_collection *aws_profile_collection_new_from_merge( struct aws_allocator *allocator, const struct aws_profile_collection *config_profiles, const struct aws_profile_collection *credentials_profiles);

LibAwsSdkutils.aws_profile_collection_release — Method

aws_profile_collection_release(collection)

Decrements a profile collection's ref count. When the ref count drops to zero, the collection will be destroyed. Returns NULL.

Prototype

struct aws_profile_collection *aws_profile_collection_release(struct aws_profile_collection *collection);

LibAwsSdkutils.aws_profile_get_name — Method

aws_profile_get_name(profile)

Returns a reference to the name of the provided profile

Prototype

const struct aws_string *aws_profile_get_name(const struct aws_profile *profile);

LibAwsSdkutils.aws_profile_get_property — Method

aws_profile_get_property(profile, property_name)

Retrieves a reference to a property with the specified name, if it exists, from a profile

Prototype

const struct aws_profile_property *aws_profile_get_property( const struct aws_profile *profile, const struct aws_string *property_name);

LibAwsSdkutils.aws_profile_get_property_count — Method

aws_profile_get_property_count(profile)

Returns how many properties a profile holds

Prototype

size_t aws_profile_get_property_count(const struct aws_profile *profile);

LibAwsSdkutils.aws_profile_property_get_sub_property — Method

aws_profile_property_get_sub_property(property, sub_property_name)

Returns a reference to the value of a sub property with the given name, if it exists, in the property

Prototype

const struct aws_string *aws_profile_property_get_sub_property( const struct aws_profile_property *property, const struct aws_string *sub_property_name);

LibAwsSdkutils.aws_profile_property_get_sub_property_count — Method

aws_profile_property_get_sub_property_count(property)

Returns how many sub properties the property holds

Prototype

size_t aws_profile_property_get_sub_property_count(const struct aws_profile_property *property);

LibAwsSdkutils.aws_profile_property_get_value — Method

aws_profile_property_get_value(property)

Returns a reference to the property's string value

Prototype

const struct aws_string *aws_profile_property_get_value(const struct aws_profile_property *property);

LibAwsSdkutils.aws_resource_name_init_from_cur — Method

aws_resource_name_init_from_cur(arn, input)

Given an ARN "Amazon Resource Name" represented as an in memory a structure representing the parts

Prototype

int aws_resource_name_init_from_cur(struct aws_resource_name *arn, const struct aws_byte_cursor *input);

LibAwsSdkutils.aws_resource_name_length — Method

aws_resource_name_length(arn, size)

Calculates the space needed to write an ARN to a byte buf

Prototype

int aws_resource_name_length(const struct aws_resource_name *arn, size_t *size);

LibAwsSdkutils.aws_sdkutils_library_clean_up — Method

aws_sdkutils_library_clean_up()

Documentation not found.

Prototype

void aws_sdkutils_library_clean_up(void);

LibAwsSdkutils.aws_sdkutils_library_init — Method

aws_sdkutils_library_init(allocator)

Documentation not found.

Prototype

void aws_sdkutils_library_init(struct aws_allocator *allocator);